chore: tabs ui updates
