+++ app/client/src/pages/Editor/IDE/EditorPane/components/SegmentedHeader.tsx
@@ -1,25 +1,15 @@
-import React, { useCallback } from "react";
-import { Button, Flex, SegmentedControl, Tooltip } from "design-system";
-import {
-  createMessage,
-  EDITOR_PANE_TEXTS,
-  MAXIMIZE_BUTTON_TOOLTIP,
-} from "@appsmith/constants/messages";
-import {
-  EditorEntityTab,
-  EditorViewMode,
-} from "@appsmith/entities/IDE/constants";
+import React from "react";
+import { Button, Flex, SegmentedControl } from "design-system";
+import { createMessage, EDITOR_PANE_TEXTS } from "@appsmith/constants/messages";
+import { EditorEntityTab } from "@appsmith/entities/IDE/constants";
 import history from "utils/history";
 import { globalAddURL } from "@appsmith/RouteBuilder";
-import { useDispatch, useSelector } from "react-redux";
+import { useSelector } from "react-redux";
 import { getCurrentPageId } from "@appsmith/selectors/entitiesSelector";
 import { useCurrentEditorState, useSegmentNavigation } from "../../hooks";
 import styled from "styled-components";
 import { useFeatureFlag } from "utils/hooks/useFeatureFlag";
 import { FEATURE_FLAG } from "@appsmith/entities/FeatureFlag";
-import { getIDEViewMode, getIsSideBySideEnabled } from "selectors/ideSelectors";
-import { setIdeEditorViewMode } from "actions/ideActions";
-import AnalyticsUtil from "../../../../../utils/AnalyticsUtil";
 
 const Container = styled(Flex)`
   #editor-pane-segment-control {
@@ -33,24 +23,15 @@ const Container = styled(Flex)`
 `;
 
 const SegmentedHeader = () => {
-  const dispatch = useDispatch();
   const isGlobalAddPaneEnabled = useFeatureFlag(
     FEATURE_FLAG.release_global_add_pane_enabled,
   );
-  const isSideBySideEnabled = useSelector(getIsSideBySideEnabled);
-  const editorMode = useSelector(getIDEViewMode);
   const pageId = useSelector(getCurrentPageId);
   const onAddButtonClick = () => {
     history.push(globalAddURL({ pageId }));
   };
   const { segment } = useCurrentEditorState();
   const { onSegmentChange } = useSegmentNavigation();
-  const handleMaximizeButtonClick = useCallback(() => {
-    AnalyticsUtil.logEvent("EDITOR_MODE_CHANGE", {
-      to: EditorViewMode.FullScreen,
-    });
-    dispatch(setIdeEditorViewMode(EditorViewMode.FullScreen));
-  }, []);
 
   return (
     <Container
@@ -89,19 +70,6 @@ const SegmentedHeader = () => {
           size="sm"
           startIcon="add-line"
         />
-      ) : null}
-      {isSideBySideEnabled &&
-      editorMode === EditorViewMode.SplitScreen &&
-      segment !== EditorEntityTab.UI ? (
-        <Tooltip content={createMessage(MAXIMIZE_BUTTON_TOOLTIP)}>
-          <Button
-            id="editor-mode-maximize"
-            isIconButton
-            kind="tertiary"
-            onClick={handleMaximizeButtonClick}
-            startIcon="maximize-v3"
-          />
-        </Tooltip>
       ) : null}
     </Container>
   );
