+++ app/client/src/pages/Editor/IDE/EditorTabs/FullScreenTabs.tsx
@@ -57,6 +57,7 @@ const FullScreenTabs = () => {
         placement="bottomRight"
       >
         <Button
+          className="ml-auto"
           id="editor-mode-minimize"
           isIconButton
           kind="tertiary"
