+++ app/client/src/pages/Editor/IDE/EditorTabs/ListButton.tsx
@@ -31,7 +31,7 @@ const ListButton = (props: Props) => {
           kind="tertiary"
           onClick={() => setOpen(true)}
         >
-          <Text kind="action-m">{items.length}</Text>
+          <Text kind="action-m">+ {items.length}</Text>
         </Button>
       </MenuTrigger>
       <MenuContent
@@ -39,6 +39,7 @@ const ListButton = (props: Props) => {
         data-testId={"t--page-selection"}
         height={items.length <= 6 ? "fit-content" : "186px"}
         side={"bottom"}
+        width="216px"
       >
         {items.map((item) => (
           <MenuItem key={item.key} onClick={() => navigateToTab(item)}>
@@ -46,7 +47,6 @@ const ListButton = (props: Props) => {
               alignItems="center"
               className={"text-ellipsis whitespace-nowrap overflow-hidden"}
               gap="spaces-2"
-              width="10rem"
             >
               <ListIconContainer>{item.icon}</ListIconContainer>
               <TabTextContainer>{item.title}</TabTextContainer>
