+++ app/client/src/pages/Editor/IDE/EditorTabs/StyledComponents.tsx
@@ -8,12 +8,13 @@ import { DEFAULT_SPLIT_SCREEN_WIDTH } from "constants/AppConstants";
  * 4px  tabs + add icon container left padding
  * 4px  tabs + add icon container right padding
  * 4px  gap between tabs and add icon
- * 4px  gap between every tabs * 4 (since max tab count is 5,
+ * 16px 4px gap between every tabs * 4 (since max tab count is 5,
  *      there will be 5 gaps)
  * 26px Add button width
  * 62px show more list button(considering 3 digit width as max)
+ * 24px width of maximize button
  * ======================================
- * 116px
+ * 151px
  *
  */
 export const StyledTab = styled(Flex)`
@@ -29,7 +30,7 @@ export const StyledTab = styled(Flex)`
   border-top-right-radius: var(--ads-v2-border-radius);
   align-items: center;
   justify-content: center;
-  max-width: calc((${DEFAULT_SPLIT_SCREEN_WIDTH} - 116px) / 5);
+  max-width: calc((${DEFAULT_SPLIT_SCREEN_WIDTH} - 151px) / 5);
 
   // After element - the seperator in between tabs
   &:not(&.active):not(:has(+ .active)):not(:last-child):after {
